requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-community
python-dotenv
SQLAlchemy
requests
numexpr
pytest